tests/wpt/web-platform-tests/2dcontext/fill-and-stroke-styles/2d.pattern.basic.type.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.pattern.basic.type</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>2d.pattern.basic.type</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

_assertDifferent(window.CanvasPattern, undefined, "window.CanvasPattern", "undefined");

window.CanvasPattern.prototype.thisImplementsCanvasPattern = true;

var img = document.getElementById('green.png');
var pattern = ctx.createPattern(img, 'no-repeat');
_assert(pattern.thisImplementsCanvasPattern, "pattern.thisImplementsCanvasPattern");


});
</script>
<img src="/images/green.png" id="green.png" class="resource">
